Improved the treeview behaviour on autofocus and opening a different form.
